config_polybar
The date block is set to open calcurse in an lxterminal window (which in turn is set to be floating and appear in a specific position in the i3 config file). The music note icon on the right opens mocp in an lxterminal window (lxterminal windows are set to be floating in i3 config).
Then theres a module to capture info of music playing in mocp and buttons to control the player. You can scroll on the information area to move backwards/forwards in the song.